The program now links as a TS6 server.
Add `sid` to the uplink configuration section.
Add a very basic IRC parser. I did a lot of benchmarking, and this came
out to be the fastest. I may change it later, but it seems good.
Add lots of TS6 handler methods.
Add a Server class for keeping track of servers on the network.

I'm not exceedingly happy with the structure things are taking. I need
to find a clean way to implement extensions/modules, and global things. I
also need to rework the socket stuff, because I'm completely unhappy with
cool.io. So those things are next on the list.