Add Gemfiles for automated services and those who know what they are
Add .travis.yml for running tests on travis-ci.org

I'm including Gemfile for things like travis-ci, and possibly other
services that use bundler for dependency resolution. I'm not actually
requiring bundler in the app itself though, and for now I'm keeping
my home-grown dependency resolution. This might change in the future,
but I feel that bundler is a dev tool, not a user tool. In the
meantime, the presence of the Gemfiles will allow anyone that knows
what bundler is to take full advantage of it. It's especially good
for running on jruby.