Major refactor for the mode stacking code, done the Right Way
Mode stacking is now totally decoupled from the protocol stuff. The way
it was done before was pretty much because I had been hacking on it
trying to get it to work, and it wasn't exactly pretty. Now all the mode
stacking itself is taken care of by the ModeStacker class, which is a total
black box with the only exposed method being ModeStacker#stack_modes.